refactor(cli): move validateTemplateName to shared util
Extracts template name validation out of the init command into
src/utils/templateName.ts so migrate no longer imports from a sibling
command. Also renames the migrate override test block to reflect that
it covers all override flags, not just --name.
