+++ packages/cli/src/commands/template/init.ts
@@ -7,6 +7,7 @@ import * as path from 'path'
 import { pathOption } from 'src/options'
 import { getRoot } from 'src/utils/filesystem'
 import { asPrimary } from 'src/utils/format'
+import { validateTemplateName } from 'src/utils/templateName'
 import {
   generateAndWriteTemplateFiles,
   GeneratedFiles,
@@ -133,24 +134,6 @@ async function addPackageJsonScripts(
     console.warn(
       '\n⚠️  Could not add scripts to package.json:',
       err instanceof Error ? err.message : err
-    )
-  }
-}
-
-/**
- * Check template name format
- */
-function templateNameRegex(name: string): boolean {
-  return /^[a-z0-9][a-z0-9_-]*[a-z0-9]$|^[a-z0-9]$/.test(name)
-}
-
-export function validateTemplateName(name: string) {
-  if (!name || name.trim().length === 0) {
-    throw new Error('Template name cannot be empty')
-  }
-  if (!templateNameRegex(name.trim())) {
-    throw new Error(
-      'Template name must contain only lowercase letters, numbers, hyphens, and underscores, and cannot start or end with a hyphen or underscore'
     )
   }
 }
